ChooseRoom input validation finished (still needs discussion with Jonathan) Split up BookingForm into two forms (BookingForm & CustomerForm)
