+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/CreateBookingServiceImpl.java
@@ -4,6 +4,7 @@ import at.fhv.hotelsoftware.application.api.CreateBookingService;
 import at.fhv.hotelsoftware.domain.api.BookingRepository;
 import at.fhv.hotelsoftware.domain.model.*;
 import at.fhv.hotelsoftware.view.form.BookingForm;
+import at.fhv.hotelsoftware.view.form.CustomerForm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -29,16 +30,15 @@ public class CreateBookingServiceImpl implements CreateBookingService {
     }
 
     @Transactional
-    public void createBooking(BookingForm bookingForm){
+    public void createBooking(BookingForm bookingForm, CustomerForm customerForm){
 
 
 
         //TODO: Input validation (later not this sprint)
         Booking booking = Booking.builder().
                 withBookingId(new BookingId(UUID.randomUUID())).
-                withCustomer(bookingForm.getFname() + " " + bookingForm.getLname()).
-                withVoucherCode(new VoucherCode(bookingForm.getVoucherCode())).
                 withCancellationDeadLine(null).
+                withCustomer(customerForm.getFname() + " " + customerForm.getLname()).
                 withBookingStatus(BookingStatus.PENDING).
                 withCheckInDate(LocalDate.parse(bookingForm.getCheckInDate())).
                 withCheckOutDate(LocalDate.parse(bookingForm.getCheckOutDate())).
