+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/api/CreateBookingService.java
@@ -2,10 +2,11 @@ package at.fhv.hotelsoftware.application.api;
 
 import at.fhv.hotelsoftware.domain.model.BookingId;
 import at.fhv.hotelsoftware.view.form.BookingForm;
+import at.fhv.hotelsoftware.view.form.CustomerForm;
 
 public interface CreateBookingService{
 
-   void createBooking(BookingForm bookingForm);
+   void createBooking(BookingForm bookingForm, CustomerForm customerForm);
