+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/view/BookingController.java
@@ -5,6 +5,7 @@ import at.fhv.hotelsoftware.application.api.ViewBookingService;
 import at.fhv.hotelsoftware.application.dto.BookingDTO;
 import at.fhv.hotelsoftware.domain.model.Booking;
 import at.fhv.hotelsoftware.view.form.BookingForm;
+import at.fhv.hotelsoftware.view.form.CustomerForm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -42,10 +43,10 @@ public class BookingController {
     @GetMapping(DASHBOARD_URL)
     public ModelAndView showDashboard(Model model) {
         List<BookingDTO> listOfBookings = viewBookingService.findTodaysCheckIns();
-        model.addAttribute("bookings", listOfBookings);
+        model.addAttribute("checkIns", listOfBookings);
 
         List<BookingDTO> listOfCheckouts = viewBookingService.findTodaysCheckOuts();
-        model.addAttribute("checkouts", listOfCheckouts);
+        model.addAttribute("checkOuts", listOfCheckouts);
 
         return new ModelAndView("dashboard");
     }
@@ -54,39 +55,44 @@ public class BookingController {
     @GetMapping(CREATE_CUSTOMER_URL)
     public ModelAndView showCustomerForm(Model model) {
         BookingForm bookingForm = new BookingForm();
+        CustomerForm customerForm = new CustomerForm();
         model.addAttribute("bookingForm", bookingForm);
+        model.addAttribute("customerForm", customerForm);
         return new ModelAndView("createCustomer");
     }
 
     @PostMapping(CHOOSE_ROOM_URL)
-    public ModelAndView submitChooseRoom(@ModelAttribute("bookingForm") @Valid BookingForm bookingForm, BindingResult result, Model model) {
+    public ModelAndView submitChooseRoom(@ModelAttribute("customerForm") @Valid CustomerForm customerForm, BindingResult result,  @ModelAttribute("bookingForm") BookingForm bookingForm, Model model) {
         if (result.hasErrors()) {
             return new ModelAndView("createCustomer");
         }
 
         model.addAttribute("bookingForm", bookingForm);
+        model.addAttribute("customerForm", customerForm);
         return new ModelAndView("chooseRoom");
     }
 
     @PostMapping(EXTRA_SERVICE_URL)
-    public ModelAndView submitExtraService(@ModelAttribute("bookingForm") @Valid BookingForm bookingForm, BindingResult result, Model model) {
+    public ModelAndView submitExtraService(@ModelAttribute("customerForm") @Valid CustomerForm customerForm, @ModelAttribute("bookingForm") @Valid BookingForm bookingForm, BindingResult result, Model model) {
         if (result.hasErrors()) {
             return new ModelAndView("chooseRoom");
         }
 
         model.addAttribute("bookingForm", bookingForm);
+        model.addAttribute("customerForm", customerForm);
         return new ModelAndView("extraService");
     }
 
     @PostMapping(BOOKING_SUMMARY_URL)
-    public ModelAndView submitBookingSummary(@ModelAttribute("bookingForm") BookingForm bookingForm, Model model) {
+    public ModelAndView submitBookingSummary(@ModelAttribute("customerForm") @Valid CustomerForm customerForm, @ModelAttribute("bookingForm") BookingForm bookingForm, Model model) {
         model.addAttribute("bookingForm", bookingForm);
+        model.addAttribute("customerForm", customerForm);
         return new ModelAndView("bookingSummary");
     }
 
     @PostMapping(WRITE_BOOKING_IN_DB)
-    public ModelAndView writeBookingInDatabase(@ModelAttribute("bookingForm") BookingForm bookingForm, Model model) {
-        createBookingService.createBooking(bookingForm);
+    public ModelAndView writeBookingInDatabase(@ModelAttribute("customerForm") @Valid CustomerForm customerForm, @ModelAttribute("bookingForm") BookingForm bookingForm, Model model) {
+        createBookingService.createBooking(bookingForm, customerForm);
         return new ModelAndView("redirect:"+"/");
     }
 }
