+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/view/annotations/DateNotNullOrEarlierThanToday.java
@@ -1,0 +1,21 @@
+package at.fhv.hotelsoftware.view.annotations;
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.FIELD, ElementType.PARAMETER})
+@Constraint(validatedBy = DateValidatorImpl.class)
+public @interface DateNotNullOrEarlierThanToday {
+
+
+    String message() default "Illegal Date";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+}
