+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/view/annotations/DateValidator.java
@@ -1,0 +1,6 @@
+package at.fhv.hotelsoftware.view.annotations;
+
+import java.lang.annotation.Annotation;
+
+public interface DateValidator extends Annotation {
+}
