+++ hotelsoftware/src/main/resources/templates/bookingSummary.html
@@ -41,7 +41,7 @@
 
     <div class="w-full bg-white m-40 p-16 rounded border border-gray-400">
 
-<form th:action="@{/writeBookingInDatabase}" th:object="${bookingForm}" method="post">
+<form th:action="@{/writeBookingInDatabase}" th:object="${customerForm}" method="post">
 
 
 
@@ -158,19 +158,19 @@
 
             <div class="w-1/3">
                 <h2 class="mb-2 font-semibold">Single </h2>
-                <input readonly type="number" id="single" class="border-2 border-gray-400 w-16" th:field="*{singleRoomCount}">
+                <input readonly type="number" id="single" class="border-2 border-gray-400 w-16" th:field="${bookingForm.singleRoomCount}">
                 <label for="single" class="mt-2 block text-sm text-gray-500">Room Count</label>
             </div>
 
             <div class="w-1/3">
                 <h2 class="mb-2 font-semibold">Double</h2>
-                <input readonly type="number" id="double" class="border-2 border-gray-400 w-16" th:field="*{doubleRoomCount}">
+                <input readonly type="number" id="double" class="border-2 border-gray-400 w-16" th:field="${bookingForm.doubleRoomCount}">
                 <label for="double" class="mt-2 block text-sm text-gray-500">Room Count</label>
             </div>
 
             <div class="w-1/3">
                 <h2 class="mb-2 font-semibold">Luxus</h2>
-                <input readonly type="number" id="luxus" class="border-2 border-gray-400 w-16" th:field="*{luxusRoomCount}">
+                <input readonly type="number" id="luxus" class="border-2 border-gray-400 w-16" th:field="${bookingForm.luxusRoomCount}">
                 <label for="luxus" class="mt-2 block text-sm text-gray-500">Room Count</label>
             </div>
 
@@ -185,12 +185,12 @@
         <div class="flex row"> <!-- Form Line -->
 
             <div class="w-1/2">
-                <input readonly type="date" id="checkInDate" class="p-2 border-2 border-gray-400 mb-2" th:field="*{checkInDate}">
+                <input readonly type="date" id="checkInDate" class="p-2 border-2 border-gray-400 mb-2" th:field="${bookingForm.checkInDate}">
                 <label for="checkInDate" class="block text-sm text-gray-500">Check-In Date</label>
             </div>
 
             <div class="w-1/2">
-                <input readonly type="date" id="checkOutDate" class="p-2 border-2 border-gray-400 mb-2" th:field="*{checkOutDate}">
+                <input readonly type="date" id="checkOutDate" class="p-2 border-2 border-gray-400 mb-2" th:field="${bookingForm.checkOutDate}">
                 <label for="checkOutDate" class="block text-sm text-gray-500">Check-Out</label>
             </div>
