+++ hotelsoftware/src/main/resources/templates/chooseRoom.html
@@ -47,17 +47,17 @@
 
     <div class="w-full bg-white m-40 p-16 rounded border border-gray-400">
 
-      <form th:action="@{/extraService}" th:object="${bookingForm}" method="post">
+      <form th:action="@{/extraService}" th:object="${bookingForm}"  method="post">
 
-        <input type="hidden" id="fname" name="fname" th:field="*{fname}"/>
-        <input type="hidden" id="lname" name="lname" th:field="*{lname}"/>
-        <input type="hidden" id="voucher" name="voucher" th:field="*{voucherCode}"/>
-        <input type="hidden" id="streetAdr" name="streetAdr" th:field="*{streetAdr}"/>
-        <input type="hidden" id="zip" name="zip" th:field="*{zip}"/>
-        <input type="hidden" id="city" name="city" th:field="*{city}"/>
-        <input type="hidden" id="country" name="country" th:field="*{country}"/>
-        <input type="hidden" id="phone" name="phone" th:field="*{phoneNumber}"/>
-        <input type="hidden" id="email" name="email" th:field="*{email}"/>
+        <input type="hidden" id="fname" name="fname" th:field="${customerForm.fname}"/>
+        <input type="hidden" id="lname" name="lname" th:field="${customerForm.lname}"/>
+        <input type="hidden" id="voucher" name="voucher" th:field="${customerForm.voucherCode}"/>
+        <input type="hidden" id="streetAdr" name="streetAdr" th:field="${customerForm.streetAdr}"/>
+        <input type="hidden" id="zip" name="zip" th:field="${customerForm.zip}"/>
+        <input type="hidden" id="city" name="city" th:field="${customerForm.city}"/>
+        <input type="hidden" id="country" name="country" th:field="${customerForm.country}"/>
+        <input type="hidden" id="phone" name="phone" th:field="${customerForm.phoneNumber}"/>
+        <input type="hidden" id="email" name="email" th:field="${customerForm.email}"/>
 
         <div class="p-4 border border-gray-300"> <!-- Logical Block -->
 
@@ -69,21 +69,33 @@
               <h2 class="mb-2 font-semibold">Single </h2>
               <input type="number" id="single" class="border-2 border-gray-400 w-16" th:field="*{singleRoomCount}">
               <label for="single" class="mt-2 block text-sm text-gray-500">Room Count</label>
-              <div class="w-4/6 p-1 font-bold text-red-700 bg-yellow-300 border border-blue-300" th:if="${#fields.hasErrors('singleRoomCount')}" th:errors="*{singleRoomCount}"></div>
+              <div class="flex rounded content-start mt-1 w-4/6 p-1 text-xs text-gray-600 bg-red-200 border-red-200 border" th:if="${#fields.hasErrors('singleRoomCount')}">
+                <svg xmlns="http://www.w3.org/2000/svg" class="pr-1" width="18" height="18" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
+                <div  th:errors="*{singleRoomCount}"></div>
+              </div>
             </div>
 
             <div class="w-1/3">
               <h2 class="mb-2 font-semibold">Double</h2>
               <input type="number" id="double" class="border-2 border-gray-400 w-16" th:field="*{doubleRoomCount}">
               <label for="double" class="mt-2 block text-sm text-gray-500">Room Count</label>
-              <div class="w-4/6 p-1 font-bold text-red-700 bg-yellow-300 border border-blue-300" th:if="${#fields.hasErrors('doubleRoomCount')}" th:errors="*{doubleRoomCount}"></div>
+              <div class="flex rounded content-start mt-1 w-4/6 p-1 text-xs text-gray-600 bg-red-200 border-red-200 border" th:if="${#fields.hasErrors('doubleRoomCount')}">
+                <svg xmlns="http://www.w3.org/2000/svg" class="pr-1" width="18" height="18" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
+                <div  th:errors="*{doubleRoomCount}"></div>
+              </div>
             </div>
 
             <div class="w-1/3">
               <h2 class="mb-2 font-semibold">Luxus</h2>
               <input type="number" id="luxus" class="border-2 border-gray-400 w-16" th:field="*{luxusRoomCount}">
-              <div class="w-4/6 p-1 font-bold text-red-700 bg-yellow-300 border border-blue-300" th:if="${#fields.hasErrors('luxusRoomCount')}" th:errors="*{luxusRoomCount}"></div>
               <label for="luxus" class="mt-2 block text-sm text-gray-500">Room Count</label>
+              <div class="flex rounded content-start mt-1 w-4/6 p-1 text-xs text-gray-600 bg-red-200 border-red-200 border" th:if="${#fields.hasErrors('luxusRoomCount')}">
+                <svg xmlns="http://www.w3.org/2000/svg" class="pr-1" width="18" height="18" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
+                <div  th:errors="*{luxusRoomCount}"></div>
+              </div>
             </div>
 
           </div>
@@ -99,11 +111,21 @@
           <div class="w-1/2">
             <input type="date" id="checkInDate" class="p-2 border-2 border-gray-400 mb-2" th:field="*{checkInDate}">
             <label for="checkInDate" class="block text-sm text-gray-500">Check-In Date</label>
+            <div class="flex rounded content-start mt-1 w-4/6 p-1 text-xs text-gray-600 bg-red-200 border-red-200 border" th:if="${#fields.hasErrors('checkInDate')}">
+              <svg xmlns="http://www.w3.org/2000/svg" class="pr-1" width="18" height="18" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
+              <div  th:errors="*{checkInDate}"></div>
+            </div>
           </div>
 
           <div class="w-1/2">
             <input type="date" id="checkOutDate" class="p-2 border-2 border-gray-400 mb-2" th:field="*{checkOutDate}">
             <label for="checkOutDate" class="block text-sm text-gray-500">Check-Out Date</label>
+            <div class="flex rounded content-start mt-1 w-4/6 p-1 text-xs text-gray-600 bg-red-200 border-red-200 border" th:if="${#fields.hasErrors('checkOutDate')}">
+              <svg xmlns="http://www.w3.org/2000/svg" class="pr-1" width="18" height="18" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
+              <div  th:errors="*{checkOutDate}"></div>
+            </div>
           </div>
 
           </div>
